Move bank_tiles.h to common directory
