+++ fasm-plugin/fasm.cc
@@ -24,7 +24,7 @@
  *   annotations on the design cells.
  */
 
-#include "../bank_tiles.h"
+#include "../common/bank_tiles.h"
 #include "kernel/log.h"
 #include "kernel/register.h"
 #include "kernel/rtlil.h"
